+++ backend/app/Api/System/Custom/js.php
@@ -1,0 +1,14 @@
+<?php
+use MythicalDash\App;
+use MythicalDash\Config\ConfigInterface;
+$router->add('/api/system/custom.js', function () {
+	App::init();
+    $appInstance = App::getInstance(true);
+    $config = $appInstance->getConfig();
+
+    $customJs = $config->getSetting(ConfigInterface::CUSTOM_JS, '');
+
+    header('Content-Type: application/javascript');
+    
+	echo $customJs;
+});