+++ backend/storage/migrations/2025-05-06.10.19-allow-longer-settings.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `mythicaldash_settings` CHANGE `value` `value` LONGTEXT CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NOT NULL COMMENT 'The value of the setting!';